organized files
move files to proper folder for organization.
