+++ app.js
@@ -1,4 +1,4 @@
-const ArrayParser = require('./arrayParser');
+const ArrayParser = require('./parser/arrayParser');
 const parserApp = new ArrayParser();
 const input = "['1a3',[null,false,['11',[112233],{easy : ['hello', {a:'a'}, 'world']},112],55, '99'],{a:'str', b:[912,[5656,33],{key : 'innervalue', newkeys: [1,2,3,4,5]}]}, true]"
