+++ parser/helper/stack.js
